+++ includes/class-tmsm-admin-cleanup.php
@@ -166,6 +166,7 @@ class Tmsm_Admin_Cleanup {
 		$this->loader->add_filter( 'oceanwp_news_enabled', $plugin_admin, 'oceanwp_news_enabled', 10 );
 		$this->loader->add_action( 'wp_dashboard_setup', $plugin_admin, 'woocommerce_remove_dashboard_widgets', 10 );
 		$this->loader->add_filter( 'admin_body_class', $plugin_admin, 'admin_body_class_role', 10, 1 );
+		$this->loader->add_filter( 'admin_email_check_interval', $plugin_admin, 'admin_email_check_interval', 10, 1 );
 
 		// Jetpack
 		$this->loader->add_filter( 'jetpack_just_in_time_msgs', $plugin_admin, 'jetpack_just_in_time_msgs');